Create FBSDKApplicationObserving protocol to decouple a few classes from FBSDKApplicationDelegate
Summary:
Currently, `BridgeAPI`, `AppLinks` and a few other components are tightly coupled with `FBSDKApplicationDelegate` for application delegate callbacks, to decouple those classes, we need to generalize the application callback logic, so `FBSDKApplicationDelegate` doesn't have to depends on those components.
- Create protocol for application delegate callbacks
- Provide functions to register and unregister application delegate callback observers

In coming diffs, I'll create separate observers for coupled components of `FBSDKApplicationDelegate`, and register them with `FBSDKApplicationDelegate`.

Reviewed By: codytwinton

Differential Revision: D14333614

fbshipit-source-id: 22284aff81b7bfaca85f02ed1ef25e9e2307b8d9